# docs/swagger.yaml
definitions:
...
  handlers.BookingRequest:
    properties:
      booking_type:
        $ref: '#/definitions/models.BookingType'
      customer_id:
        type: integer
      date:
        type: string
      duration:
        type: integer
      rider_count:
        type: integer
      start_time:
        type: string
      track_id:
        type: integer
    type: object
...
  models.BookingType:
    enum:
    - Обычное бронирование
    - Бронирование всего трека
    - Бронирование всего картодрома
    type: string
    x-enum-varnames:
    - CommonBooking
    - OneTrackBooking
    - AllTrackBooking
...
  models.Kart:
    properties:
      id:
        type: integer
      kartModel:
        $ref: '#/definitions/models.KartModel'
      kartModelID:
        type: integer
      kartodrom:
        $ref: '#/definitions/models.Kartodrom'
      kartodromID:
        type: integer
      status:
        $ref: '#/definitions/models.KartStatus'
    type: object
  models.KartModel:
    properties:
      category:
        $ref: '#/definitions/models.KartModelStatus'
      desc:
        type: string
      id:
        type: integer
      maxHeight:
        type: integer
      maxSpeed:
        type: integer
      maxWeight:
        type: integer
      name:
        type: string
      pricePerMin:
        type: number
    type: object
  models.KartModelStatus:
    enum:
    - детский
    - обычный
    - спортивный
    - гоночный
    - электрический
    type: string
    x-enum-varnames:
    - KidKart
    - CommonKart
    - SportKart
    - RaceKart
    - ElectoKart
  models.KartStatus:
    enum:
    - Доступен
    - В использовании
    - Сломан
    - Недоступен
    type: string
    x-enum-varnames:
    - Available
    - InUse
    - Broken
    - InStopList
  models.Kartodrom:
    properties:
      city:
        type: string
      email:
        type: string
      id:
        type: integer
      latitude:
        type: number
      location:
        type: string
      longitude:
        type: number
      name:
        type: string
      phone:
        type: string
      schedules:
        items:
          $ref: '#/definitions/models.KartodromSchedule'
        type: array
    type: object
  models.KartodromSchedule:
    properties:
      closeTime:
        type: string
      dayOfWeek:
        type: integer
      id:
        type: integer
      kartodrom:
        $ref: '#/definitions/models.Kartodrom'
      kartodromID:
        type: integer
      openTime:
        type: string
    type: object
...
  services.LoginInput:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  services.RegisterInput:
    properties:
      birthday:
        type: string
      email:
        type: string
      fname:
        type: string
      login:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
info:
  contact: {}
paths:
  /bookings:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные бронирования
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.BookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать бронирование
      tags:
      - booking
  /bookings/{id}/activate:
    post:
      parameters:
      - description: ID бронирования
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Активировать бронирование
      tags:
      - booking
  /bookings/{id}/cancel:
    post:
      parameters:
      - description: ID бронирования
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Отменить бронирование
      tags:
      - booking
  /bookings/{id}/close:
    post:
      parameters:
      - description: ID бронирования
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Завершить бронирование
      tags:
      - booking
...
  /kartodrom/{id}/free-karts:
    get:
      parameters:
      - description: ID картодрома
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Kart'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить доступные карты в картодроме
      tags:
      - kart
...
  /karts/{id}/book:
    post:
      parameters:
      - description: ID карта
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Забронировать карт
      tags:
      - kart
  /karts/{id}/broken:
    post:
      parameters:
      - description: ID карта
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Пометить карт как сломанный
      tags:
      - kart
  /karts/{id}/free:
    post:
      parameters:
      - description: ID карта
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Освободить карт
      tags:
      - kart
  /karts/{id}/repair:
    post:
      parameters:
      - description: ID карта
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Восстановить карт
      tags:
      - kart
  /karts/{id}/stoplist:
    post:
      parameters:
      - description: ID карта
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить карт в стоп-лист
      tags:
      - kart
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/services.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Логин пользователя
      tags:
      - auth
...
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/services.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - auth
...
swagger: "2.0"